+++ packages/utils/src/exports.ts
@@ -26,6 +26,7 @@ import { waitForEvent } from './waitForEvent'
 import { TimeoutError, withTimeout } from './withTimeout'
 import { composeAbortSignals } from './composeAbortSignals'
 import { waitForCondition } from './waitForCondition'
+import { withRateLimit } from './withRateLimit'
 
 export {
     BrandedString,
@@ -52,6 +53,7 @@ export {
     wait,
     waitForCondition,
     waitForEvent,
+    withRateLimit,
     withTimeout,
 }